Update on combat system along with some fixes regarding projectiles Updated magic system to be using kotlin and refactored a little

+++ src/com/rs/java/game/player/CombatDefinitions.java
@@ -53,7 +53,7 @@ public final class CombatDefinitions implements Serializable {
 	private boolean defensiveCasting;
 	public transient boolean instantAttack;
 	private transient boolean dungeonneringSpellBook;
-	public int spellBook;
+	public byte spellBook;
 	private byte autoCastSpell;
 
 	public int getSpellId() {
